ollama-examples/requirements.txt
ollama>=0.6.0
pandas>=2.3.0
python-dotenv>=1.0.0
python-dateutil>=2.9.0
requests>=2.32.0